remove recipes feature
more details:

creating a gem just for a single line of deploy
script is overkill and not agile. Therefore, now
removing all traces of the recipe feature.

This makes allows me to remove a lot of code and
finally implement the multistage deploy feature,
as we can just add hooks for different rake tasks
throughout deployment.